Add client-side expiry filtering and auto-refresh to pending samples
